quick mock/requirements.unstable.txt
beautifulsoup
boto
django-celery-with-redis
django<1.6
django-allauth
django-analytical>=0.12.1
django-annoying
django_compressor
django-heroku-memcacheify
django-heroku-postgresify
django-extensions
django-nose
django-selector
django-staticfiles
django-storages
django-ses
fabric
gunicorn
requests
pillow
psycopg2
python-firebase
sorl-thumbnail
south
splinter
versiontools

# Dev only
django-debug-toolbar
ipython